Remap keys like in my XTerm configuration.
In XTerm, I find it useful to remap keys as shown in the following
excerpt from my .Xresources

xterm*vt100.translations: #override \
! ...
!   C-. -> C-^
    Ctrl <KeyPress> period: string(0x1e) \n\
!   C-, -> C-_
    Ctrl <KeyPress> comma: string(0x1f) \n\
!   C-i -> C-]
    Ctrl <KeyPress> i: string(0x1d) \n\
!   C-m -> C-\
    Ctrl <KeyPress> m: string(0x1c) \n\

The reason the first two remappings are useful is that they allow me
to type C-^ and C-_ using the middle and ring fingers on my left hand
respectively (I use a Dvorak layout).  Without the remapping, both
"key chords" are on right pinky (and require me to use the Shift key).

The reason remapping C-i and C-m is useful is that, in a terminal, the
<Tab> key already sends C-i, and the <Enter> key already sends C-m.
Also, C-] and C-\ are more awkward to type on a Dvorak keyboard than
C-i and C-m.

Having a way to effect the remappings above allows me to switch
completely from XTerm to Terminology.